Change the type of the 'amount' long to int.

+++ src/main/java/com/ttarum/member/dto/request/CartUpdateRequest.java
@@ -8,5 +8,5 @@ import lombok.Getter;
 public class CartUpdateRequest {
 
     private final long itemId;
-    private final long amount;
+    private final int amount;
 }
